Mocked in translation of label fields on collection and manifest
